Create .nvmrc
With this file in the project folder, we can keep track of the used node version and also have automation in places that can use that bit of information e.g. to change automatically the node version like this in your rc file of the terminal

# This checks is a .nvmrc is givin and if so, it loads it
if [ -f .nvmrc ]
  then
  nvm use
fi

wrote that into my .zshrc file, works nice and smooth. 🤓